Moving files to their right place
