+++ fire_atlas_editor/lib/screens/open_screen/open_screen.dart
@@ -8,7 +8,7 @@ import '../../services/storage.dart';
 import '../../widgets/icon_button.dart';
 import '../../widgets/button.dart';
 import '../../widgets/container.dart';
-import '../../widgets/scaffold.dart';
+import '../widgets/scaffold.dart';
 import '../../widgets/text.dart';
 import '../../utils/select_file.dart';
