+++ fire_atlas_editor/lib/screens/widgets/scaffold.dart
